requirements.txt
beautifulsoup4==4.14.3
dnspython==2.8.0
Flask==3.1.2
flask_dance==7.1.0
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
fpdf==1.7.2
joblib==1.5.1
joblib==1.5.2
keras==3.11.3
keras==3.12.0
nltk==3.9.1
nltk==3.9.2
numpy==2.3.5
pandas==2.3.3
pdfminer3k==1.3.4
pdfminer_six==20250506
pdfplumber==0.11.7
pefile==2024.8.26
Pillow==12.0.0
plotly==6.2.0
plotly==6.4.0
PyPDF2==3.0.1
python-dotenv==1.2.1
Requests==2.32.5
scipy==1.16.3
selenium==4.39.0
SQLAlchemy==2.0.45
tensorflow==2.20.0
urllib3==2.6.2
Werkzeug==3.1.4
whois==1.20240129.2
WTForms==3.2.1
xhtml2pdf==0.2.17
yara==1.7.7
yara_python==4.5.2
